Changed type of input variable

+++ modules/calc_module.cpp
@@ -1,6 +1,6 @@
 #include "calc_module.h"
 
-float pi_count(int x)
+float pi_count(float x)
 {
 	return x*3.14;
 }
